+++ src/examples/refuse_post_example.c
@@ -39,7 +39,7 @@ ahc_echo (void *cls,
           const char *url,
           const char *method,
           const char *version,
-          const char *upload_data, unsigned int *upload_data_size, void **ptr)
+          const char *upload_data, size_t *upload_data_size, void **ptr)
 {
   static int aptr;
   const char *me = cls;
